Add function to switch between scenes
