+++ src/main/resources/view/MatchTestPage.fxml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.VBox?>
+
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.layout.StackPane?>
+
+<VBox xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1">
+    <SplitPane id="testMSplitPane" fx:id="testMSplitPane" dividerPositions="0.5" VBox.vgrow="ALWAYS">
+        <VBox fx:id="testMQuestionListView" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="testMQuestionListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+
+        <VBox fx:id="testMAnswerListView" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="testMAnswerListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+    </SplitPane>
+</VBox>
